requirements.txt
alembic==1.7.6
arrow==1.2.2
backports.zoneinfo==0.2.1
click==8.0.4
dateparser==1.1.0
ecdsa==0.17.0
Flask==2.0.3
Flask-Cors==3.0.10
Flask-Migrate==2.7.0
Flask-Moment==1.0.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.11.2
importlib-resources==5.4.0
itsdangerous==2.1.1
Jinja2==3.0.3
Mako==1.2.0
MarkupSafe==2.1.0
moment==0.12.1
psycopg2==2.9.3
psycopg2-binary==2.9.1
pyasn1==0.4.8
python-dateutil==2.8.2
python-jose==3.3.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.2
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.32
times==0.7
tzdata==2021.5
tzlocal==4.1
Werkzeug==2.0.3
zipp==3.7.0
